fix: address CodeRabbit review for PR #58
- pyproject.toml: remove duplicate tactile-bridge entry point
- __init__.py: sort __all__ alphabetically (package + bridge)
- CI: ubuntu-24.04 → ubuntu-latest
- integration_demo.py: replace broad except AttributeError with getattr check
- hand_zenoh_bridge.py: joint/control_mode dtype int32 → uint16
- tactile_bridge.py: atomic frame snapshot (single get_tactile_raw + local normalize)
- tactile_bridge.py: use raw serial_number in capability, _bridge_id for keys only
- touch_board.hpp/cpp: usb_pid int32_t → uint16_t
- transport.hpp/usb.cpp: usb_pid int32_t → uint16_t
- usb.cpp: track detached interfaces, reattach symmetrically in destructor

+++ src/wujihandpy/bridge/hand_zenoh_bridge.py
@@ -779,7 +779,7 @@ class HandBridge:
         # Other writes need SDO access
         with self._hand_lock:
             if path == "joint/control_mode":
-                self.hand.write_joint_control_mode(np.array(value, dtype=np.int32))
+                self.hand.write_joint_control_mode(np.array(value, dtype=np.uint16))
             elif path == "joint/enabled":
                 self.hand.write_joint_enabled(np.array(value, dtype=bool))
             elif path == "joint/effort_limit":
