+++ wujihandcpp/src/transport/transport.hpp
@@ -29,10 +29,10 @@ public:
 };
 
 std::unique_ptr<ITransport>
-    create_usb_transport(uint16_t usb_vid, int32_t usb_pid, const char* serial_number);
+    create_usb_transport(uint16_t usb_vid, uint16_t usb_pid, const char* serial_number);
 
 std::unique_ptr<ITransport>
-    create_usb_transport(uint16_t usb_vid, int32_t usb_pid, const char* serial_number,
+    create_usb_transport(uint16_t usb_vid, uint16_t usb_pid, const char* serial_number,
                          int interface_num, unsigned char in_ep, unsigned char out_ep);
 
 } // namespace wujihandcpp::transport
